requirements.txt
flask==3.0.0
line-bot-sdk==3.5.1
python-dotenv==1.0.0
openai==0.28.1
redis==5.0.1
langdetect==1.0.9
deep-translator==1.11.4
requests==2.31.0
pytz==2023.3.post1
httpx>=0.23.0,<1.0.0
googlemaps==4.10.0
gunicorn==21.2.0
gevent==23.9.1
Werkzeug==3.0.1
